requirements.txt
Flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
scikit-learn==1.3.0
gunicorn==21.2.0
openai[datalib]
numpy